Refactor runner -> pipeline
