+++ src/quartic/__init__.py
@@ -3,5 +3,5 @@ from .ipython_magics import *
 from .quartic import Quartic
 from .utils import QuarticException
 
-from .runner.step import step
-from .runner.dataset import writer
+from .pipeline.step import step
+from .pipeline.dataset import writer
