+++ src/quartic/pipeline/README.md
