+++ src/quartic/pipeline/cli.py
