+++ src/quartic/pipeline/exceptions.py
